+++ pkg/config/snmpmetriccfg.go
@@ -4,22 +4,24 @@ import (
 	"errors"
 	"fmt"
 	"github.com/Knetic/govaluate"
-
+	"regexp"
+	"strconv"
 	"strings"
 )
 
 //SnmpMetricCfg Metric config
 type SnmpMetricCfg struct {
-	ID          string  `xorm:"'id' unique" binding:"Required"` //name of the key in the config array
-	FieldName   string  `xorm:"field_name" binding:"Required"`
-	Description string  `xorm:"description"`
-	BaseOID     string  `xorm:"baseoid"`
-	DataSrcType string  `xorm:"datasrctype"`
-	GetRate     bool    `xorm:"getrate"` //ony Valid with COUNTERS
-	Scale       float64 `xorm:"scale"`
-	Shift       float64 `xorm:"shift"`
-	IsTag       bool    `xorm:"'istag' default 0"`
-	ExtraData   string  `xorm:"extradata"` //Only Valid with STRINGPARSER and STRINGEVAL
+	ID          string         `xorm:"'id' unique" binding:"Required"` //name of the key in the config array
+	FieldName   string         `xorm:"field_name" binding:"Required"`
+	Description string         `xorm:"description"`
+	BaseOID     string         `xorm:"baseoid"`
+	DataSrcType string         `xorm:"datasrctype"`
+	GetRate     bool           `xorm:"getrate"` //ony Valid with COUNTERS
+	Scale       float64        `xorm:"scale"`
+	Shift       float64        `xorm:"shift"`
+	IsTag       bool           `xorm:"'istag' default 0"`
+	ExtraData   string         `xorm:"extradata"`  //Only Valid with STRINGPARSER, STRINGEVAL , BITS , BITSCHK
+	Names       map[int]string `xorm:"-" json:"-"` //BitString Name array
 }
 
 /*
@@ -47,7 +49,7 @@ func (m *SnmpMetricCfg) Init() error {
 	case "Counter64", "COUNTER64": //raw and Cooked increment of Counter64
 	case "COUNTERXX": //raw and Coocked increment with non_negative behaivour of Counters
 	case "TimeTicks", "TIMETICKS": //raw and cooked to second of timeticks
-	case "BITS":
+	case "BITS", "BITSCHK":
 	case "OCTETSTRING":
 	case "OID":
 	case "HWADDR":
@@ -58,8 +60,27 @@ func (m *SnmpMetricCfg) Init() error {
 	default:
 		return errors.New("UnkNown DataSourceType:" + m.DataSrcType + " in metric Config " + m.ID)
 	}
-	if m.DataSrcType == "BITS" && len(m.ExtraData) == 0 {
-		return errors.New("BITS type requires extradata to work " + m.ID)
+	if m.DataSrcType == "BITSCHK" {
+		if len(m.ExtraData) == 0 {
+			return errors.New("BITSCHK type requires extradata to work " + m.ID)
+		}
+		_, err := strconv.Atoi(m.ExtraData)
+		if err != nil {
+			return errors.New("BITSCHK type requires extradata to be a positive integer to work: ERROR " + err.Error())
+		}
+	}
+	if m.DataSrcType == "BITS" {
+		if len(m.ExtraData) == 0 {
+			return errors.New("BITS type requires extradata to work " + m.ID)
+		}
+		//named bits array construction for this Config
+		re := regexp.MustCompile("([a-zA-Z0-9]+)\\(([0-9]+)\\)")
+		m.Names = make(map[int]string)
+		str := re.FindAllStringSubmatch(m.ExtraData, -1)
+		for _, x := range str {
+			i, _ := strconv.Atoi(x[2])
+			m.Names[i] = x[1]
+		}
 	}
 	if m.DataSrcType != "STRINGEVAL" && m.DataSrcType != "CONDITIONEVAL" && !strings.HasPrefix(m.BaseOID, ".") {
 		return errors.New("Bad BaseOid format:" + m.BaseOID + " in metric Config " + m.ID)